JavaCall.addClassPath in __init__
It's important that `JavaCall.addClassPath` be invoked in your `__init__` so that it gets invoked upon `using View5D`. Other Java packages for Julia can also add to the classpath in their `__init__`. Putting it in the bare module does not work, since it would only be invoked at compile time.